+++ src/main/groovy/com/getkeepsafe/dexcount/OutputFormat.groovy
@@ -23,20 +23,26 @@ public enum OutputFormat {
     /**
      * Specifies that method counts will be printed in a flat list of packages.
      */
-    LIST,
+    LIST('.txt'),
 
     /**
      * Specifies that the output will be pretty-printed as an tree.
      */
-    TREE,
+    TREE('.txt'),
 
     /**
      * Specifies that the output will be a pretty-printed JSON object.
      */
-    JSON,
+    JSON('.json'),
 
     /**
      * Specifies that output will be a YAML document.
      */
-    YAML,
+    YAML('.yml');
+
+    public String extension
+
+    OutputFormat(String extension) {
+        this.extension = extension
+    }
 }