+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.4.1 (unreleased)
 ----------
 
+* Change output-file extensions to '.json' and '.yml' for JSON and YAML outputs.
 * Revert pull request #32 and remove `@Input` from `apkOrDex` (issues #35 and #46)
 
 0.4.0 (released 23 January 2016)