+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ----------
 
 * Change output-file extensions to '.json' and '.yml' for JSON and YAML outputs.
-* Revert pull request #32 and remove `@Input` from `apkOrDex` (issues #35 and #46)
+* Revert pull request #32 and remove `@Input` from `apkOrDex` (issues #40 and #46)
 
 0.4.0 (released 23 January 2016)
 ----------